+++ R/init.R
@@ -1,7 +1,7 @@
 #' @importFrom graphics image.default image par plot title box text axis plot.new plot.window rasterImage layout lcm contour hist
 #' @importFrom grDevices dev.capabilities dev.size grey rgb col2rgb
 #' @importFrom utils head tail setTxtProgressBar txtProgressBar packageVersion
-#' @importFrom stats na.omit runif aggregate setNames predict quantile var
+#' @importFrom stats na.omit runif aggregate setNames predict quantile var complete.cases
 #' @importFrom tools file_ext
 #' @importFrom methods as slotNames new slot
 #' @importFrom abind abind adrop asub